cstring bug修复。解决部分警告。
cstring:规范内存管理调用，增加移动构造和复制构造函数，修复几处错误的内存释放判断和_append中错误的复制长度。
Keil *.uvoptx中的tvExp值代表工程目录某个文件分组是否展开，提交前请保持全部收起。
解决了下列文件中部分类型转换相关警告：
Microsecond.hpp,NamedVariant.hpp,QNum.hpp,
构造函数增加explicit限定:
QNum,cstring

+++ bsp/stm32/stm32f429-st-nucleo/utility/Microsecond.hpp
@@ -125,7 +125,7 @@ namespace Microsecond
     timestamp_t us_tick(void)
     {
         //TODO:std::chrono测试验证，API统一
-        return std::chrono::microseconds().count();
+        return (uint32_t)std::chrono::microseconds().count();
     }
 #endif
